+++ dev/pack_and_release.sh
@@ -1,16 +1,19 @@
 #!/usr/bin/env bash
 
-rm out/*.vsix
+set -e
+
+if (( $# !=  1 )); then
+	>&2 echo "Must have 1 argument: patch, minor or major"
+	exit 1
+fi
+
+tag=$(npm version $1)
+
+echo "$tag"
+
+rm -f out/*.vsix
 vsce pack -o out
 
-ver=$(jq '.version' \
-      --raw-output \
-      -- < package.json)
-
-tag="v$ver"
-echo $tag
-
-git tag -a $tag
 git push --follow-tags
 
-gh release create $tag --notes-from-tag --fail-on-no-commits ./out/*.vsix
+gh release create $tag --fail-on-no-commits ./out/*.vsix