+++ CHANGELOG.md
@@ -5,6 +5,10 @@ All notable changes to this project will be documented in this file.
 The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.1.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
+## [Unreleased]
+### Changed
+- Made the pack_and_release.sh script use npm version
+
 ## [[0.0.2]](https://github.com/cookie1170/unity-asmdef/releases/tag/v0.0.2) - 18 Jan 2026
 ### Added
 - Publisher metadata to package.json